Added unit test file for my user stories
Please use this unit test file as an example to create tests for each of your assigned user stories. You will have to either add cases to the 'test.ged' file to trigger the case in your user story, or create a new file with that case and specifiy it in the test (and also add it to the repo).